Change: rename TicketCatalogueService packages to lowercase
