--- TicketCatalogueService/src/main/kotlin/it/polito/wa2/ticketcatalogueservice/Repositories/TicketRepository.kt
@@ -1,12 +1,0 @@
-package it.polito.wa2.ticketcatalogueservice.Repositories
-
-import it.polito.wa2.ticketcatalogueservice.Entities.Order
-import it.polito.wa2.ticketcatalogueservice.Entities.Ticket
-import kotlinx.coroutines.flow.Flow
-import org.springframework.data.r2dbc.repository.Query
-import org.springframework.data.repository.kotlin.CoroutineCrudRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface TicketRepository: CoroutineCrudRepository<Ticket, Long>{
-}
