+++ TicketCatalogueService/src/main/kotlin/it/polito/wa2/ticketcatalogueservice/entities/Order.kt
@@ -1,4 +1,4 @@
-package it.polito.wa2.ticketcatalogueservice.Entities
+package it.polito.wa2.ticketcatalogueservice.entities
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Column
@@ -6,27 +6,27 @@ import org.springframework.data.relational.core.mapping.Table
 
 @Table("orders")
 data class Order(
-    @Id
+        @Id
     val id: Long,
 
-    @Column("ticketid")
+        @Column("ticketid")
     val ticketId: Long,
 
-    val quantity: Int,
+        val quantity: Int,
 
-    @Column("paymentid")
+        @Column("paymentid")
     val paymentId: Long,
 
-    @Column("userid")
+        @Column("userid")
     val userId: Long,
 
-    @org.springframework.data.annotation.Transient
+        @org.springframework.data.annotation.Transient
     val ticket: Ticket,
 
-    @org.springframework.data.annotation.Transient
+        @org.springframework.data.annotation.Transient
     val user: User,
 
-    @org.springframework.data.annotation.Transient
+        @org.springframework.data.annotation.Transient
     val paymentInformation: PaymentInformation
 
 )
