+++ TicketCatalogueService/src/main/kotlin/it/polito/wa2/ticketcatalogueservice/repositories/TicketRepository.kt
@@ -1,0 +1,9 @@
+package it.polito.wa2.ticketcatalogueservice.repositories
+
+import it.polito.wa2.ticketcatalogueservice.entities.Ticket
+import org.springframework.data.repository.kotlin.CoroutineCrudRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface TicketRepository: CoroutineCrudRepository<Ticket, Long>{
+}
