+++ TicketCatalogueService/src/main/kotlin/it/polito/wa2/ticketcatalogueservice/security/JwtAuthorizationFilter.kt
@@ -1,4 +1,4 @@
-package it.polito.wa2.ticketcatalogueservice.Security
+package it.polito.wa2.ticketcatalogueservice.security
 
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.authority.SimpleGrantedAuthority
